refactor(metrics): split args_valid metric
